backend/requirements.txt
django>=5.1,<6
djangorestframework>=3.15
django-cors-headers>=4.3
django-encrypted-model-fields>=0.6
django-celery-beat>=2.6
celery>=5.3
redis>=5.0
python-dotenv>=1.0
requests>=2.31
anthropic>=0.25
google-api-python-client>=2.118
serpapi>=0.1
psycopg2-binary>=2.9
